requirements.txt
interactions
gptzero
requests
python-dotenv

